requirements.txt
pandas
numpy
openai
tqdm
scikit-learn
streamlit
python-dotenv




























